# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NCS Warehouse Management API",
        "description": "API untuk manajemen gudang PT. NCS Bandung dengan algoritma Simulated Annealing",
        "contact": {
            "name": "NCS Development Team",
            "email": "dev@ncs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/kategori-barang": {
            "get": {
                "tags": [
                    "Kategori Barang"
                ],
                "summary": "Daftar kategori barang",
                "description": "Mengambil daftar kategori barang dengan pagination dan filter",
                "operationId": "getKategoriBarang",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pencarian berdasarkan nama atau kode kategori",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "aktif",
                        "in": "query",
                        "description": "Filter berdasarkan status aktif (1/0)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kategori barang berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar kategori barang berhasil diambil"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/KategoriBarang"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Kategori Barang"
                ],
                "summary": "Buat kategori barang baru",
                "description": "Membuat kategori barang baru dengan validasi data",
                "operationId": "storeKategoriBarang",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_kategori",
                                    "kode_kategori"
                                ],
                                "properties": {
                                    "nama_kategori": {
                                        "type": "string",
                                        "example": "Elektronik"
                                    },
                                    "kode_kategori": {
                                        "type": "string",
                                        "example": "ELK"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Kategori untuk barang elektronik"
                                    },
                                    "aktif": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategori barang berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori barang berhasil dibuat"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/KategoriBarang"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/kategori-barang/{id}/toggle-status": {
            "patch": {
                "tags": [
                    "Kategori Barang"
                ],
                "summary": "Toggle status aktif kategori barang",
                "description": "Mengubah status aktif/nonaktif kategori barang",
                "operationId": "toggleKategoriBarangStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kategori barang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status kategori barang berhasil diubah",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Status kategori barang berhasil diubah"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/KategoriBarang"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kategori barang tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/kategori-barang/stats": {
            "get": {
                "tags": [
                    "Kategori Barang"
                ],
                "summary": "Statistik kategori barang",
                "description": "Mengambil statistik kategori barang",
                "operationId": "getKategoriBarangStats",
                "responses": {
                    "200": {
                        "description": "Statistik kategori barang berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Statistik kategori barang berhasil diambil"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_kategori": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "kategori_aktif": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "kategori_nonaktif": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "total_barang_per_kategori": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "kategori_terpopuler": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "nama_kategori": {
                                                                "type": "string"
                                                            },
                                                            "jumlah_barang": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/kategori-barang/aktif": {
            "get": {
                "tags": [
                    "Kategori Barang"
                ],
                "summary": "Daftar kategori barang aktif",
                "description": "Mengambil daftar kategori barang yang aktif untuk dropdown",
                "operationId": "getActiveKategoriBarang",
                "responses": {
                    "200": {
                        "description": "Daftar kategori barang aktif berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar kategori barang aktif berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "nama_kategori": {
                                                        "type": "string"
                                                    },
                                                    "kode_kategori": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "KategoriBarang": {
                "title": "Kategori Barang",
                "description": "Model kategori barang",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_kategori": {
                        "type": "string",
                        "example": "Elektronik"
                    },
                    "kode_kategori": {
                        "type": "string",
                        "example": "ELK"
                    },
                    "deskripsi": {
                        "type": "string",
                        "example": "Kategori untuk barang elektronik"
                    },
                    "aktif": {
                        "type": "boolean",
                        "example": true
                    },
                    "jumlah_barang": {
                        "type": "integer",
                        "example": 25
                    },
                    "status_text": {
                        "type": "string",
                        "example": "Aktif"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}